requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
contourpy==1.1.0
cycler==0.11.0
dgl==1.1.1
filelock==3.12.2
fonttools==4.41.1
idna==3.4
imageio==2.31.1
importlib-resources==6.0.0
Jinja2==3.1.2
joblib==1.3.1
kiwisolver==1.4.4
lazy_loader==0.3
llvmlite==0.40.1
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
networkx==2.8
numba==0.57.1
numpy==1.24.4
opencv-python==4.8.0.74
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
psutil==5.9.5
pynndescent==0.5.10
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
requests==2.31.0
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
six==1.16.0
sympy==1.12
threadpoolctl==3.2.0
tifffile==2023.7.18
torch==2.0.1
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
umap-learn==0.5.3
urllib3==2.0.4
zipp==3.16.2
